backend/requirements.txt
altgraph==0.17.5
asgiref==3.11.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
clr_loader==0.2.10
contourpy==1.3.3
cycler==0.12.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
fonttools==4.61.1
gunicorn==25.0.2
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.2
packaging==26.0
pandas==3.0.0
pefile==2024.8.26
pillow==12.1.0
pycparser==3.0
pygame==2.6.1
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
pyparsing==3.3.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.18.0
python-dateutil==2.9.0.post0
pythonnet==3.0.5
pywin32-ctypes==0.2.3
reportlab==4.4.9
requests==2.32.5
setuptools==80.10.2
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.11.0
